Split basics.mjs
